+++ apmconnector/logs.go
@@ -1,7 +1,6 @@
 package apmconnector
 
 import (
-	"fmt"
 	"go.opentelemetry.io/collector/pdata/plog"
 	"go.opentelemetry.io/collector/pdata/ptrace"
 )
